requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.3
Django==3.1.2
django-embed-video==1.3.3
djangorestframework==3.12.2
idna==2.10
isort==5.6.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
pylint==2.6.0
python-resize-image==1.1.19
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
